+++ app/models/concerns/user_helper.rb
@@ -22,29 +22,34 @@ module UserHelper
     ).all
   end
 
-  def assign_as(role_name, to:)
-    unless to.respond_to? :journal
-      fail NoMethodError <<-ERROR.strip_heredoc
-        Sorry, I don't know which journal you want this role to apply to.
-        #{to.class.name} needs a :journal method.
-      ERROR
-    end
+  def assign_to!(assigned_to:, role:)
+    role = get_role_for_thing(assigned_to, role)
+    Assignment.where(
+      user: self,
+      role: role,
+      assigned_to: assigned_to
+    ).first_or_create!
+  end
 
-    role = role_for_assignment(role_name, to.journal)
-    assignments.create(role: role, assigned_to: to)
+  def resign_from!(assigned_to:, role:)
+    role = get_role_for_thing(assigned_to, role)
+    assignments.where(role: role, assigned_to: assigned_to).destroy_all
   end
 
   private
 
-  def role_for_assignment(role_name, journal)
-    role = journal.roles.find_by(name: role_name) ||
-           Role.find_by(name: role_name, journal: nil)
+  # Return the role with the name `role_name` associated with a given thing.
+  # Return role if role is already a Role.
+  def get_role_for_thing(thing, role_name)
+    return role_name if role_name.is_a?(Role)
 
-    return role if role
-
-    fail ActiveRecord::RecordNotFound <<-ERROR.strip_heredoc
-        Sorry, I couldn't find a role #{role_name} in #{journal},
-        or with no journal at all.
+    # role_name is a string, need to get the right role for the journal
+    journal = thing.is_a?(Journal) ? thing : thing.try(:journal)
+    unless journal
+      fail <<-ERROR.strip_heredoc
+        Expected #{thing} to be a journal or respond to journal method
       ERROR
+    end
+    Role.find_by!(journal: journal, name: role_name)
   end
 end