+++ engines/tahi_standard_tasks/client/app/adapters/data-availability-task.js
@@ -1,0 +1,5 @@
+import TaskAdapter from 'tahi/adapters/task';
+
+let DataAvailabilityTaskAdapter = TaskAdapter.extend();
+
+export default DataAvailabilityTaskAdapter;