+++ engines/tahi_standard_tasks/spec/models/send_to_apex_task_spec.rb
@@ -5,11 +5,12 @@ describe TahiStandardTasks::SendToApexTask do
     FactoryGirl.create :paper, :with_tasks
   end
   let!(:task) do
-    TahiStandardTasks::SendToApexTask.create!(
-      paper: paper,
-      old_role: 'editor',
-      phase: paper.phases.first
-    )
+    FactoryGirl.create(:send_to_apex_task, paper: paper)
+  end
+
+  describe '.restore_defaults' do
+    include_examples '<Task class>.restore_defaults update title to the default'
+    include_examples '<Task class>.restore_defaults update old_role to the default'
   end
 
   describe '#apex_deliveries association' do