+++ CHANGELOG.md
@@ -4,6 +4,9 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 
 ## [{version}] - {release_date}
 ### Added
+- Ability for invited reviewer to provide decline reason and suggested
+  reviewers when declining invitation.
+- Ability to remove user roles via CSV file.
 ### Changed
 ### Deprecated
 ### Removed